Add length in given dimension as property.
Added new procedure that counts a node's children in a given dimension and stores the coung in a property 'length_<dimension>' on the node.

Define length in glossary.

Renamed 'leafDimension' to 'childDimension' throughout the AggregatingProcedures since it's not necessarily about leaves.

Use new stringifiers and debugging feature of the 'convenience-matchers'.
Added some tests.